+++ includes/classes/class-taxonomies.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * LSX Tour Operator - Taxonomy Class
+ * Tour Operator - Taxonomy Class
  *
  * @package   lsx
  * @author    LightSpeed
@@ -36,8 +36,8 @@ class Taxonomies extends Frame {
 	 * @param array  $config The taxonomy config arguments.
 	 */
 	protected function register_object( $slug, $config ) {
+		$config['args'] = apply_filters( 'lsx_to_register_taxonomy_args', $config['args'], $slug );
 		register_taxonomy( $slug, $config['object_types'], $config['args'] );
 		$this->object[ $slug ] = get_taxonomy( $slug );
 	}
-
 }