+++ includes/classes/class-tour-operator.php
@@ -6,6 +6,7 @@ use lsx\admin\Admin;
 use lsx\admin\Pages;
 use lsx\admin\Settings;
 use lsx\admin\Setup;
+use lsx\admin\Permalinks;
 use lsx\blocks\Bindings;
 use lsx\blocks\Patterns;
 use lsx\blocks\Registration;
@@ -13,7 +14,7 @@ use lsx\blocks\Templates;
 use lsx\integrations\facetwp\Post_Connections;
 
 /**
- * LSX Tour Operator Main Class
+ * Tour Operator Main Class
  *
  * @package   tour_operator
  * @author    LightSpeed
@@ -72,6 +73,14 @@ class Tour_Operator {
 	public $pages;
 
 	/**
+	 * Holds the Permalinks instance.
+	 *
+	 * @since   1.1.0
+	 * @var     \lsx\admin\Permalinks
+	 */
+	public $permalinks;
+
+	/**
 	 * Holds the Taxonomies instance.
 	 *
 	 * @since   1.1.0
@@ -95,7 +104,7 @@ class Tour_Operator {
 	public $legacy;
 
 	/**
-	 * LSX Tour Operator constructor.
+	 * Tour Operator constructor.
 	 */
 	public function __construct() {
 		// init legacy.
@@ -217,6 +226,7 @@ class Tour_Operator {
 	public function setup() {
 		require_once( LSX_TO_PATH . 'vendor/content-models/create-content-model.php' );
 
+		$this->permalinks = new Permalinks();
 		$this->pages      = Pages::init();
 		$this->taxonomies = Taxonomies::init();
 		$this->admin      = new Admin();
@@ -226,8 +236,6 @@ class Tour_Operator {
 		$this->registration = new Registration();
 		$this->patterns   = new Patterns();
 		$this->templates   = new Templates();
-
-		do_action( 'qm/debug', $this->templates );
 
 		// Files that wont load with the badly written spl_autoregister function.
 		require_once LSX_TO_PATH . 'includes/classes/class-post-connections.php';