+++ includes/partials/settings.php
@@ -3,7 +3,7 @@
 $settings_pages = tour_operator()->legacy->get_post_types();
 ?>
 <div class="wrap lsx-to-settings">
-	<h1><?php echo esc_html__( 'LSX Tour Operator Settings', 'tour-operator' ); ?></h1>
+	<h1><?php echo esc_html__( 'Tour Operator Settings', 'tour-operator' ); ?></h1>
 	<form method="post">
 		<?php
 		include( LSX_TO_PATH . 'includes/partials/navigation.php' );