+++ changelog.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [[2.0.2]](https://github.com/lightspeeddevelopment/tour-operator/releases/tag/2.0.2) - 2025-05-06
+
+### Added
+- The Itinerary Included and Excluded field handling and block output.
+- `lsx_to_wetu_map_url_params` filter to allow 3rd party plugins to change the WETU map attributes and what shows.
+- Permalink Settings fields, to allow the URL change for Travel Styles, Accommodation Type, and Brands.
+- The user can now alter the taxonomy slugs using the "Permalink" settings in WordPress.
+
+### Fixed
+- Fixed the output of the Special Interests - [BH-77](https://www.bugherd.com/projects/430995/tasks/77)
+- Drinks Basis labels outputting "Drinks Basis" - [BH-80](https://www.bugherd.com/projects/430995/tasks/80)
+- The Archive redirects and single redirects now use the correct post type arg parameters to disable the relevant option. #367
+- The version number display on the welcome page. - #482
+
+### Enhancements Added
+- `lsx_to_wetu_map_url_params` filter to allow 3rd party plugins to change the WETU map attributes. [7364e48
+](https://github.com/lightspeedwp/tour-operator/commit/7364e48dbbe8ac9130a6de1afc3eb4a63e1937f4)
+- Adding in a "none" option for the "Special Interests" and the "Friendly" custom field selections. [edd55f3
+](https://github.com/lightspeedwp/tour-operator/commit/edd55f3ccb06e285146db23d3e21df25cc66920b)
+
+### Additional
+- Removed commented out code.
+- Removed unused template tags.
+- Changed the label "LSX Tour Operator" to "Tour Operator" in all files.
+
+### Security
+- Tested with WordPress 6.8.1
+
 ## [[2.0.1]](https://github.com/lightspeeddevelopment/tour-operator/releases/tag/2.0.1) - 2025-01-24
 
 ### Added
@@ -37,7 +65,7 @@
 - **Regions Section**: Created a carousel to display child regions in the Single Country template. Hidden if no regions are associated. [#384](https://github.com/lightspeedwp/tour-operator/issues/384)
 
 #### Plugin Code
-- **New Templates**: Updated Single Post templates for Tours, Accommodation, Destinations, and more to match the LSX Tour Operator Design System. [#334](https://github.com/lightspeedwp/tour-operator/issues/334)
+- **New Templates**: Updated Single Post templates for Tours, Accommodation, Destinations, and more to match the Tour Operator Design System. [#334](https://github.com/lightspeedwp/tour-operator/issues/334)
 - **Custom Query Loops**: Introduced variations for dynamically displaying related content (e.g., related tours, accommodation) based on specific criteria. [#431](https://github.com/lightspeedwp/tour-operator/issues/431)
 - **Custom Block Dashicons**: Added a custom compass icon for branding consistency. [#401](https://github.com/lightspeedwp/tour-operator/issues/401)